ported to Qt6.4.2

+++ ScriptCommunicator/colorWidgets/color_utils.cpp
@@ -24,10 +24,10 @@
 namespace color_widgets {
 namespace detail {
 
-QColor color_from_lch(qreal hue, qreal chroma, qreal luma, qreal alpha )
+QColor color_from_lch(float hue, float chroma, float luma, float alpha )
 {
-    qreal h1 = hue*6;
-    qreal x = chroma*(1-qAbs(std::fmod(h1,2)-1));
+    float h1 = hue*6;
+    float x = chroma*(1-qAbs(std::fmod(h1,2)-1));
     QColor col;
     if ( h1 >= 0 && h1 < 1 )
         col = QColor::fromRgbF(chroma,x,0);
@@ -42,7 +42,7 @@ QColor color_from_lch(qreal hue, qreal chroma, qreal luma, qreal alpha )
     else if ( h1 < 6 )
         col = QColor::fromRgbF(chroma,0,x);
 
-    qreal m = luma - color_lumaF(col);
+    float m = luma - color_lumaF(col);
 
     return QColor::fromRgbF(
         qBound(0.0,col.redF()+m,1.0),
@@ -51,11 +51,11 @@ QColor color_from_lch(qreal hue, qreal chroma, qreal luma, qreal alpha )
         alpha);
 }
 
-QColor color_from_hsl(qreal hue, qreal sat, qreal lig, qreal alpha )
+QColor color_from_hsl(float hue, float sat, float lig, float alpha )
 {
-    qreal chroma = (1 - qAbs(2*lig-1))*sat;
-    qreal h1 = hue*6;
-    qreal x = chroma*(1-qAbs(std::fmod(h1,2)-1));
+    float chroma = (1 - qAbs(2*lig-1))*sat;
+    float h1 = hue*6;
+    float x = chroma*(1-qAbs(std::fmod(h1,2)-1));
     QColor col;
     if ( h1 >= 0 && h1 < 1 )
         col = QColor::fromRgbF(chroma,x,0);
@@ -70,7 +70,7 @@ QColor color_from_hsl(qreal hue, qreal sat, qreal lig, qreal alpha )
     else if ( h1 < 6 )
         col = QColor::fromRgbF(chroma,0,x);
 
-    qreal m = lig-chroma/2;
+    float m = lig-chroma/2;
 
     return QColor::fromRgbF(
         qBound(0.0,col.redF()+m,1.0),
